+++ src/main/java/SuperDelivery/service/idm/models/LocationLatLon.java
@@ -3,36 +3,38 @@ package SuperDelivery.service.idm.models;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
+import java.math.BigDecimal;
+
 @JsonPropertyOrder({"lat", "lon"})
 public class LocationLatLon {
     @JsonProperty(value = "lat")
-    private float lat;
+    private BigDecimal lat;
     @JsonProperty(value = "lon")
-    private float lon;
+    private BigDecimal lon;
 
     private LocationLatLon(LocationLatLonBuilder builder) {
         this.lat = builder.lat;
         this.lon = builder.lon;
     }
 
-    public float getLat() {
+    public BigDecimal getLat() {
         return lat;
     }
 
-    public float getLon() {
+    public BigDecimal getLon() {
         return lon;
     }
 
     public static class LocationLatLonBuilder {
-        private float lat;
-        private float lon;
+        private BigDecimal lat;
+        private BigDecimal lon;
 
-        public LocationLatLonBuilder setLat(float lat) {
+        public LocationLatLonBuilder setLat(BigDecimal lat) {
             this.lat = lat;
             return this;
         }
 
-        public LocationLatLonBuilder setLon(float lon) {
+        public LocationLatLonBuilder setLon(BigDecimal lon) {
             this.lon = lon;
             return this;
         }
